+++ leropa/parser/types.py
@@ -1,0 +1,15 @@
+"""Common type aliases for parser structures."""
+
+from __future__ import annotations
+
+SubParagraphList = list["SubParagraph"]
+ParagraphList = list["Paragraph"]
+NoteList = list["Note"]
+HistoryList = list["HistoryEntry"]
+ArticleList = list[str]
+ArticleDataList = list["Article"]
+ChapterList = list["Chapter"]
+TitleList = list["Title"]
+BookList = list["Book"]
+SectionList = list["Section"]
+SubsectionList = list["Subsection"]
