+++ leropa/parser/utils.py
@@ -1,246 +1,20 @@
-"""Parse legal documents from legislatie.just.ro."""
+"""Utility functions for parsing document structures."""
 
 from __future__ import annotations
 
 import re
-from dataclasses import asdict, dataclass, field
-from pathlib import Path
-from typing import Any, List
+from typing import Any
 
-import requests  # type: ignore[import-untyped]
-from bs4 import BeautifulSoup
-
-# Type aliases
-SubParagraphList = List["SubParagraph"]
-ParagraphList = List["Paragraph"]
-NoteList = List["Note"]
-HistoryList = List["HistoryEntry"]
-ArticleList = List[str]
-ArticleDataList = List["Article"]
-ChapterList = List["Chapter"]
-TitleList = List["Title"]
-BookList = List["Book"]
-SectionList = List["Section"]
-SubsectionList = List["Subsection"]
-
-
-@dataclass
-class SubParagraph:
-    """Lettered or numbered sub-paragraph within a paragraph.
-
-    Attributes:
-        sub_id: Identifier for the sub-paragraph element in the source HTML.
-        label: Enumerated label such as "a)" or "(1)".
-        text: Visible text content of the sub-paragraph without the label.
-    """
-
-    sub_id: str
-    label: str
-    text: str
-
-
-@dataclass
-class HistoryEntry:
-    """Version entry from the document history.
-
-    Attributes:
-        ver_id: Identifier for the document version.
-        date: Consolidation or republication date.
-    """
-
-    ver_id: str
-    date: str
-
-
-@dataclass
-class DocumentInfo:
-    """Metadata about the parsed document.
-
-    Attributes:
-        source: Source URL for the document.
-        ver_id: Identifier for the document version.
-        title: Document title from the HTML metadata.
-        description: Document description from the HTML metadata.
-        keywords: Document keywords from the HTML metadata.
-        history: Chronological list of earlier document versions.
-        prev_ver: Identifier for previous version if available.
-        next_ver: Identifier for next version if available.
-    """
-
-    source: str
-    ver_id: str
-    title: str | None = None
-    description: str | None = None
-    keywords: str | None = None
-    history: HistoryList = field(default_factory=list)
-    prev_ver: str | None = None
-    next_ver: str | None = None
-
-
-@dataclass
-class Article:
-    """Represents a single article from the document.
-
-    Attributes:
-        article_id: Identifier for the article element in the source HTML.
-        full_text: Full text content of the article.
-        paragraphs: Ordered collection of paragraphs within the article.
-        notes: Notes attached to the article body.
-    """
-
-    article_id: str
-    full_text: str
-    paragraphs: ParagraphList = field(default_factory=list)
-    notes: NoteList = field(default_factory=list)
-
-
-@dataclass
-class Paragraph:
-    """Represents a paragraph from an article.
-
-    Attributes:
-        par_id: Identifier for the paragraph element in the source HTML.
-        text: Visible text content of the paragraph without the label.
-        label: Enumerated label such as "(1)" if present.
-        subparagraphs: Ordered collection of sub-paragraphs.
-        notes: Notes with amendment information.
-    """
-
-    par_id: str
-    text: str
-    label: str | None = None
-    subparagraphs: SubParagraphList = field(default_factory=list)
-    notes: NoteList = field(default_factory=list)
-
-
-@dataclass
-class Note:
-    """Amendment note attached to an article or paragraph.
-
-    Attributes:
-        note_id: Identifier for the note element in the source HTML.
-        text: Full textual content of the note.
-        date: Date when the amendment took effect, if available.
-        subject: Portion of the document that was modified.
-        law_number: Number of the amending law.
-        law_date: Publication date of the amending law.
-        monitor_number: Number of the "Monitorul Oficial" issue.
-        monitor_date: Date of the "Monitorul Oficial" issue.
-        replaced: Text that was replaced.
-        replacement: Text that replaced the original.
-    """
-
-    note_id: str
-    text: str
-    date: str | None = None
-    subject: str | None = None
-    law_number: str | None = None
-    law_date: str | None = None
-    monitor_number: str | None = None
-    monitor_date: str | None = None
-    replaced: str | None = None
-    replacement: str | None = None
-
-
-@dataclass
-class Subsection:
-    """Subsection grouping articles inside a section.
-
-    Attributes:
-        subsection_id: Identifier for the subsection body in the source HTML.
-        title: Subsection label such as "§1".
-        description: Descriptive text for the subsection if present.
-        articles: Ordered list of article identifiers contained in the
-            subsection.
-    """
-
-    subsection_id: str
-    title: str
-    description: str | None = None
-    articles: ArticleList = field(default_factory=list)
-
-
-@dataclass
-class Section:
-    """Section grouping subsections and articles.
-
-    Attributes:
-        section_id: Identifier for the section body in the source HTML.
-        title: Section label such as "Secţiunea I".
-        description: Descriptive text for the section if present.
-        subsections: Ordered list of subsections within the section.
-        articles: Identifiers of articles that appear directly under the
-            section.
-    """
-
-    section_id: str
-    title: str
-    description: str | None = None
-    subsections: SubsectionList = field(default_factory=list)
-    articles: ArticleList = field(default_factory=list)
-
-
-@dataclass
-class Chapter:
-    """Chapter grouping sections and articles.
-
-    Attributes:
-        chapter_id: Identifier for the chapter body in the source HTML.
-        title: Chapter label such as "Capitolul I".
-        description: Descriptive text for the chapter if present.
-        sections: Ordered list of sections within the chapter.
-        articles: Ordered list of article identifiers contained in the chapter.
-    """
-
-    chapter_id: str
-    title: str
-    description: str | None = None
-    sections: SectionList = field(default_factory=list)
-    articles: ArticleList = field(default_factory=list)
-
-
-@dataclass
-class Title:
-    """Title grouping chapters, sections and articles.
-
-    Attributes:
-        title_id: Identifier for the title body in the source HTML.
-        title: Title label such as "Titlul I".
-        description: Descriptive text for the title if present.
-        chapters: Ordered list of chapters within the title.
-        sections: Ordered list of sections within the title.
-        articles: Identifiers of articles that appear directly under the title.
-    """
-
-    title_id: str
-    title: str
-    description: str | None = None
-    chapters: ChapterList = field(default_factory=list)
-    sections: SectionList = field(default_factory=list)
-    articles: ArticleList = field(default_factory=list)
-
-
-@dataclass
-class Book:
-    """Book grouping titles, chapters, sections and articles.
-
-    Attributes:
-        book_id: Identifier for the book body in the source HTML.
-        title: Book label such as "Cartea I".
-        description: Descriptive text for the book if present.
-        titles: Ordered list of titles within the book.
-        chapters: Chapters directly contained in the book.
-        sections: Sections directly contained in the book.
-        articles: Identifiers of articles that appear directly under the book.
-    """
-
-    book_id: str
-    title: str
-    description: str | None = None
-    titles: TitleList = field(default_factory=list)
-    chapters: ChapterList = field(default_factory=list)
-    sections: SectionList = field(default_factory=list)
-    articles: ArticleList = field(default_factory=list)
+from .article import Article
+from .book import Book
+from .chapter import Chapter
+from .note import Note
+from .paragraph import Paragraph
+from .section import Section
+from .sub_paragraph import SubParagraph
+from .subsection import Subsection
+from .title import Title
+from .types import NoteList, ParagraphList
 
 
 def _normalize_whitespace(text: str) -> str:
@@ -700,159 +474,3 @@ def _ensure_subsection(
         section.subsections.append(subsection)
 
     return subsection
-
-
-def parse_html(html: str, ver_id: str) -> dict[str, Any]:
-    """Parse HTML content into structured data.
-
-    Args:
-        html: Raw HTML content of the legal document.
-        ver_id: Identifier for the document version.
-
-    Returns:
-        Structured representation of the document.
-    """
-
-    soup = BeautifulSoup(html, "html.parser")
-
-    # Extract document metadata from meta tags.
-    meta_title: Any = soup.find("meta", attrs={"name": "title"})
-    description_tag: Any = soup.find("meta", attrs={"name": "description"})
-    keywords_tag: Any = soup.find("meta", attrs={"name": "keywords"})
-
-    # Fall back to the HTML title when meta title is missing.
-    title = (
-        meta_title.get("content")
-        if meta_title and meta_title.get("content")
-        else soup.title.get_text(strip=True)
-        if soup.title
-        else None
-    )
-
-    description = description_tag.get("content") if description_tag else None
-    keywords = keywords_tag.get("content") if keywords_tag else None
-
-    # Collect historical versions from the consolidation list.
-    history: HistoryList = []
-    history_div: Any = soup.find("div", id="istoric_fa")
-    if history_div:
-        # Iterate through all links representing previous versions.
-        for link in history_div.find_all("a"):
-            href = link.get("href")
-
-            # Skip entries without hyperlinks as they point to the current
-            # version.
-            if not href:
-                continue
-
-            # Extract the version identifier from the URL.
-            ver_match = re.search(r"/(\d+)(?:\?|$)", href)
-            if not ver_match:
-                continue
-
-            # Determine the date from the link text or title attribute.
-            text = link.get_text(strip=True)
-            title_text = link.get("title", "")
-            source = (
-                text if re.search(r"\d{2}\.\d{2}\.\d{4}", text) else title_text
-            )
-            date_match = re.search(r"(\d{2}\.\d{2}\.\d{4})", source)
-            if not date_match:
-                continue
-
-            history.append(
-                HistoryEntry(
-                    ver_id=ver_match.group(1), date=date_match.group(1)
-                )
-            )
-
-    # Store all parsed articles for backward compatibility.
-    parsed_articles: ArticleDataList = []
-
-    # Temporary registries for hierarchical structures to avoid duplicates.
-    books: dict[str, Book] = {}
-    titles: dict[str, Title] = {}
-    chapters: dict[str, Chapter] = {}
-    sections: dict[str, Section] = {}
-    subsections: dict[str, Subsection] = {}
-
-    for art_tag in soup.find_all("span", class_="S_ART"):
-        # Parse the article tag into a dataclass.
-        article = _parse_article(art_tag)
-        if article is None:
-            continue
-
-        # Ensure parent containers exist and retrieve them.
-        book = _ensure_book(art_tag, books)
-        title_obj = _ensure_title(art_tag, titles, book)
-        chapter = _ensure_chapter(art_tag, chapters, title_obj, book)
-        section = _ensure_section(art_tag, sections, chapter, title_obj, book)
-        subsection = _ensure_subsection(art_tag, subsections, section)
-
-        # Attach the article id to the deepest container available.
-        if subsection:
-            subsection.articles.append(article.article_id)
-        elif section:
-            section.articles.append(article.article_id)
-        elif chapter:
-            chapter.articles.append(article.article_id)
-        elif title_obj:
-            title_obj.articles.append(article.article_id)
-        elif book:
-            book.articles.append(article.article_id)
-
-        parsed_articles.append(article)
-
-    source = f"https://legislatie.just.ro/Public/DetaliiDocument/{ver_id}"
-
-    # Store metadata and history in the document info.
-    document = DocumentInfo(
-        source=source,
-        ver_id=ver_id,
-        title=title,
-        description=description,
-        keywords=keywords,
-        # Include the parsed history.
-        history=history,
-        # The latest previous version becomes prev_ver for convenience.
-        prev_ver=history[0].ver_id if history else None,
-    )
-
-    return {
-        "document": asdict(document),
-        "articles": [asdict(a) for a in parsed_articles],
-        "books": [asdict(b) for b in books.values()],
-    }
-
-
-CACHE_DIR = Path.home() / ".leropa"
-
-
-def fetch_document(
-    ver_id: str, cache_dir: Path | None = None
-) -> dict[str, Any]:
-    """Fetch document HTML, using local cache when possible.
-
-    Args:
-        ver_id: Identifier for the document version to fetch.
-        cache_dir: Directory used for caching downloaded HTML files.
-
-    Returns:
-        Parsed document structure.
-    """
-
-    cache_dir = cache_dir or CACHE_DIR
-    cache_dir.mkdir(parents=True, exist_ok=True)
-    cache_file = cache_dir / f"{ver_id}.html"
-
-    if cache_file.exists():
-        html = cache_file.read_text(encoding="utf-8")
-    else:
-        url = f"https://legislatie.just.ro/Public/DetaliiDocument/{ver_id}"
-        response = requests.get(url, timeout=30)
-        response.raise_for_status()
-
-        html = response.text
-        cache_file.write_text(html, encoding="utf-8")
-
-    return parse_html(html, ver_id)
